requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.0
langchain-openai==0.0.2.post1
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.2.2
transformers==4.37.2
torch==2.1.2
langchainhub==0.1.14
pydantic==2.6.1
requests==2.31.0
python-dotenv==1.0.0
pypdf==3.17.1
huggingface-hub==0.20.3

#langchain==0.1.0
langchain-core==0.1.23
#langchain-openai==0.0.2.post1
langchain-anthropic==0.1.1
#langchain-community==0.0.13
#chromadb==0.4.22
anthropic
#pydantic==2.6.1
uvicorn[standard]
uvicorn[standard]
